Refractoring + stations endpoint + scraping added
Got a bit carried away... I separated mav_api's stuff into it's own
module and in the meantime I set up scraping because I noticed the
legacy scraper can go down sometimes. New endpoint also added to serve
all stations, which will be much better for planning routes.
